pgcrypto: port OpenPGP armor + symmetric encrypt/decrypt + compression + key_id + native xdes
- ASCII armor/dearmor/pgp_armor_headers (CRC24 + radix-64), text[] header arrays + materialize-SRF
- pgp_sym_encrypt/decrypt: S2K (simple/salted/iterated), OpenPGP CFB (normal+resync), MDC, session keys, aes/3des/cast5/bf
- compression: ZIP (raw deflate) + ZLIB via miniz_oxide
- pgp_key_id: packet parse + v4 SHA1 fingerprint (RSA/ElGamal/DSA pubkeys, no MPI math)
- native FreeSec crypt-des port fixes crypt-xdes _/!!!!!!! case
Public-key encrypt/decrypt remain graceful-ERROR stubs.
22 unit tests green incl a known real-gpg ZIP message round-trip.

+++ crates/contrib/pgcrypto/src/pgp/s2k.rs
@@ -116,7 +116,7 @@ impl S2k {
                     md.update(key);
                 }
                 PGP_S2K_ISALTED => {
-                    let count = s2k_decode_count(self.iter) as usize;
+                    let count = s2k_decode_count(self.iter as i32) as usize;
                     md.update(&self.salt);
                     md.update(key);
                     let mut curcnt = PGP_S2K_SALT + key.len();
